requirements-tests.txt
pytest
coverage
pytest-cov
pylint
codecov
tqdm
hypothesis-networkx
